Abstract `MigrationNotification`
